+++ adapters/Moloco/CHANGELOG.md
@@ -1,10 +1,16 @@
 ## Moloco iOS Mediation Adapter Changelog
 
+#### Next version
+
+- Updated the adapter to retrieve the Moloco SDK version from the main thread.
+
 #### [Version 3.5.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/moloco/MolocoAdapter-3.5.0.0.zip)
+
 - Initial release.
 - Added bidding support for banner (includes MREC), interstitial and rewarded ad formats.
 - Verified compatibility with Moloco SDK version 3.5.0.
 
 Built and tested with:
+
 - Google Mobile Ads SDK version 11.13.0.
 - Moloco SDK version 3.5.0.